http file helper for testing
